backend/requirements.txt
fastapi
uvicorn
gspread
google-auth
python-dotenv
scapy
requests
apscheduler
